grpc_server/requirements.txt
grpcio==1.31.0
grpcio-tools==1.31.0
json-protobuf==1.0.0
protobuf==3.13.0
six==1.15.0
